% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gillespie_dm}
\alias{gillespie_dm}
\title{Gillespie's Direct Method in C++}
\usage{
gillespie_dm(M0, tmax, pars, haz, seed, info = 100L, prealloc = 10000L)
}
\arguments{
\item{M0}{the initial marking of the net}

\item{tmax}{the maximum simulation time}

\item{pars}{a named list of parameters to pass to \code{haz}}

\item{haz}{an R function that takes 3 arguments (time, state, pars) and returns a vector of hazards for each event}

\item{seed}{a positive integer seed for the random number generator (\url{http://www.cplusplus.com/reference/random/})}

\item{info}{print information on simulation every \code{info} events}

\item{prealloc}{how much memory to preallocate for output}
}
\description{
Simulate a trajectory from a stochastic petri net model.
}
\examples{
\dontrun{
library(gillespieCXX)
library(ggplot2)
library(reshape2)
data(SIR)
out <- gillespie_dm(M0 = SIR$M0,tmax = 250,pars = SIR$pars,haz = SIR$hazard,seed = 12214L,info = 100)
out <- as.data.frame(out)
ggplot(data = melt(out,id.vars="time")) +
  geom_line(aes(x=time,y=value,color=variable)) +
  theme_bw()
}

}
