% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SIR}
\alias{SIR}
\title{Stochastic SIR model SPN}
\description{
A named list that contains the SIR model expressed as a stochastic petri net (SPN). To be used with 
\code{\link[gillespieCXX]{gillespie_dm}}.
}
\keyword{data}
